Mv metadata table to view file
